Move text search tests to the appropriate module
Issue: #297
