+++ search/src/test/java/name/mlopatkin/andlogview/search/text/SearchRequestParserTest.java
@@ -12,16 +12,14 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
- *
  */
 
-package name.mlopatkin.andlogview.search;
+package name.mlopatkin.andlogview.search.text;
 
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.lenient;
 
-import name.mlopatkin.andlogview.search.text.SearchRequestParser;
+import name.mlopatkin.andlogview.search.RequestCompilationException;
 
 import org.junit.Assert;
 import org.junit.Before;
