+++ search/src/test/java/name/mlopatkin/andlogview/search/text/SearcherBuilderTest.java
@@ -12,15 +12,13 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
- *
  */
 
-package name.mlopatkin.andlogview.search;
+package name.mlopatkin.andlogview.search.text;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 
-import name.mlopatkin.andlogview.search.text.SearcherBuilder;
+import name.mlopatkin.andlogview.search.RequestCompilationException;
 
 import com.google.common.collect.ImmutableList;
